+++ tests/php/Unit/Shared/Api/ProjectIndexTest.php
@@ -26,6 +26,37 @@ final class ProjectIndexTest extends TestCase
         self::assertSame(['user\\foo', 'user\\bar'], $index->namespaces());
     }
 
+    public function test_it_reports_a_namespace_that_has_no_definitions(): void
+    {
+        $location = new Location('empty.phel', 1, 5, 1, 10);
+        $index = new ProjectIndex([], [], ['user.empty' => $location]);
+
+        self::assertSame(['user.empty'], $index->namespaces());
+        self::assertSame(1, $index->countNamespaces());
+    }
+
+    public function test_it_counts_a_namespace_once_when_its_definitions_spell_it_with_backslashes(): void
+    {
+        $index = new ProjectIndex(
+            ['user\\foo/a' => $this->makeDefinition('user\\foo', 'a')],
+            [],
+            ['user.foo' => new Location('foo.phel', 1, 5, 1, 13)],
+        );
+
+        self::assertSame(['user\\foo'], $index->namespaces());
+        self::assertSame(1, $index->countNamespaces());
+    }
+
+    public function test_it_resolves_a_namespace_declaration_location_in_either_spelling(): void
+    {
+        $location = new Location('pprint.phel', 3, 5, 3, 16);
+        $index = new ProjectIndex([], [], ['phel.pprint' => $location]);
+
+        self::assertSame($location, $index->namespaceLocation('phel.pprint'));
+        self::assertSame($location, $index->namespaceLocation('phel\\pprint'));
+        self::assertNull($index->namespaceLocation('missing.namespace'));
+    }
+
     public function test_it_filters_definitions_by_namespace(): void
     {
         $definitions = [